use concrete type ThreeThresholdAlgorithm in ClusterFromRaw; move inlines to the .h
